+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # ChangeLog
 
-## v2.0.2
-* Move file assets from the `assets` to the `ldlib2` folder
-* Fixed cross-OS platform file separator char
+## v2.0.3
+* Fixed Mixins check